Respect IBus's orientation settings
Summary: Prior to this, the default settings forces the orientation of Rime's input dialog to be horizontal regardless of what is set in "Ibus Preferences". This commit comments out the line that causes this, so that users can set the orientation directly in IBus, instead of manually changing the config files (which sometimes require root permissions).

Test Plan:
Typed a few characters, made sure that now Rime respects IBus's orientation setting.

Signed-off-by: Gavin Zhao <git@gzgz.dev>

Reviewers: #triage_team, Girtablulu

Reviewed By: #triage_team, Girtablulu

Subscribers: Girtablulu

Differential Revision: https://dev.getsol.us/D12748